Split terminal session adapter
