+++ Sources/QuillCodeApp/WorkspaceRemoteGitHunkCommandBuilder.swift
@@ -65,6 +65,6 @@ enum WorkspaceRemoteGitHunkCommandBuilder {
     }
 
     private static func shellSingleQuoted(_ value: String) -> String {
-        WorkspaceTerminalEngine.shellSingleQuoted(value)
+        WorkspaceTerminalSessionAdapter.shellSingleQuoted(value)
     }
 }
