+++ Sources/QuillCodeApp/WorkspaceRemoteGitPushCommandBuilder.swift
@@ -35,6 +35,6 @@ enum WorkspaceRemoteGitPushCommandBuilder {
     }
 
     private static func shellSingleQuoted(_ value: String) -> String {
-        WorkspaceTerminalEngine.shellSingleQuoted(value)
+        WorkspaceTerminalSessionAdapter.shellSingleQuoted(value)
     }
 }
