+++ Sources/QuillCodeApp/WorkspaceRemoteProjectToolExecutor.swift
@@ -256,6 +256,6 @@ struct WorkspaceRemoteProjectToolExecutor: Sendable, Hashable {
     }
 
     private static func shellSingleQuoted(_ value: String) -> String {
-        WorkspaceTerminalEngine.shellSingleQuoted(value)
+        WorkspaceTerminalSessionAdapter.shellSingleQuoted(value)
     }
 }
